+++ pkg/server/commands/reverseexpand/reverse_expand_weighted.go
@@ -9,6 +9,7 @@ import (
 
 	aq "github.com/emirpasic/gods/queues/arrayqueue"
 	"go.opentelemetry.io/otel/trace"
+	"go.uber.org/zap"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 	weightedGraph "github.com/openfga/language/pkg/go/graph"
@@ -17,7 +18,6 @@ import (
 	"github.com/openfga/openfga/internal/concurrency"
 	"github.com/openfga/openfga/internal/graph"
 	"github.com/openfga/openfga/internal/stack"
-	"github.com/openfga/openfga/internal/utils"
 	"github.com/openfga/openfga/internal/validation"
 	"github.com/openfga/openfga/pkg/storage"
 	"github.com/openfga/openfga/pkg/telemetry"
@@ -140,7 +140,7 @@ func (c *ReverseExpandQuery) loopOverEdges(
 		// we don't traverse the exact same edge more than once.
 		goingToUserset := toNode.GetNodeType() == weightedGraph.SpecificTypeAndRelation
 		if goingToUserset {
-			key := edge.GetFrom().GetUniqueLabel() + toNode.GetUniqueLabel()
+			key := edge.GetFrom().GetUniqueLabel() + toNode.GetUniqueLabel() + edge.GetTuplesetRelation()
 			if _, loaded := c.visitedUsersetsMap.LoadOrStore(key, struct{}{}); loaded {
 				// we've already visited this userset through this edge, exit to avoid an infinite cycle
 				continue
@@ -257,7 +257,7 @@ func (c *ReverseExpandQuery) loopOverEdges(
 				if err != nil {
 					return err
 				}
-				err = c.intersectionHandler(ctx, newReq, resultChan, intersectionEdges, sourceUserType, resolutionMetadata)
+				err = c.intersectionHandler(pool, newReq, resultChan, intersectionEdges, sourceUserType, resolutionMetadata)
 				if err != nil {
 					return err
 				}
@@ -266,7 +266,7 @@ func (c *ReverseExpandQuery) loopOverEdges(
 				if err != nil {
 					return err
 				}
-				err = c.exclusionHandler(ctx, newReq, resultChan, exclusionEdges, sourceUserType, resolutionMetadata)
+				err = c.exclusionHandler(ctx, pool, newReq, resultChan, exclusionEdges, sourceUserType, resolutionMetadata)
 				if err != nil {
 					return err
 				}
@@ -394,11 +394,6 @@ func (c *ReverseExpandQuery) executeQueryJob(
 
 	currentReq.relationStack = newStack
 
-	// Ensure that we haven't already run this query
-	if c.isDuplicateQuery(userFilter, typeRel, newStack) {
-		return nil, nil
-	}
-
 	objectType, relation := tuple.SplitObjectRelation(typeRel)
 
 	filteredIter, err := c.buildFilteredIterator(ctx, currentReq, objectType, relation, userFilter)
@@ -484,34 +479,6 @@ func buildUserFilter(
 	return []*openfgav1.ObjectRelation{filter}, nil
 }
 
-func (c *ReverseExpandQuery) isDuplicateQuery(
-	userFilter []*openfgav1.ObjectRelation,
-	typeRel string,
-	relationStack stack.Stack[typeRelEntry],
-) bool {
-	objectType, relation := tuple.SplitObjectRelation(typeRel)
-
-	// Create a unique key for the current query to avoid duplicate work.
-	key := utils.Reduce(userFilter, "", func(accumulator string, current *openfgav1.ObjectRelation) string {
-		return current.String() + accumulator
-	})
-
-	// Also need to consider the query's context, as it is possible that many different branches end at the
-	// same leaf or have a common ancestor as part of a unique traversal, and we don't want to kill a unique branch.
-	stackStr := ""
-	for relationStack != nil {
-		var val typeRelEntry
-		val, relationStack = stack.Pop(relationStack)
-		stackStr += fmt.Sprintf("%v", val)
-	}
-	key += stackStr
-
-	key += relation + objectType
-	_, loaded := c.queryDedupeMap.LoadOrStore(key, struct{}{})
-
-	return loaded
-}
-
 // buildFilteredIterator constructs the iterator used when reverse_expand queries for tuples.
 // The returned iterator MUST have .Stop() called on it.
 func (c *ReverseExpandQuery) buildFilteredIterator(
@@ -584,6 +551,61 @@ func (c *ReverseExpandQuery) findCandidatesForLowestWeightEdge(
 	})
 }
 
+// callCheckForCandidates calls check on the list objects candidate against non lowest weight edges.
+func (c *ReverseExpandQuery) callCheckForCandidate(
+	ctx context.Context,
+	req *ReverseExpandRequest,
+	tmpResult *ReverseExpandResult,
+	resultChan chan<- *ReverseExpandResult,
+	userset *openfgav1.Userset,
+	isAllowed bool,
+) error {
+	handlerFunc := c.localCheckResolver.CheckRewrite(ctx,
+		&graph.ResolveCheckRequest{
+			StoreID:              req.StoreID,
+			AuthorizationModelID: c.typesystem.GetAuthorizationModelID(),
+			TupleKey:             tuple.NewTupleKey(tmpResult.Object, req.Relation, req.User.String()),
+			ContextualTuples:     req.ContextualTuples,
+			Context:              req.Context,
+			Consistency:          req.Consistency,
+			RequestMetadata:      graph.NewCheckRequestMetadata(),
+		}, userset)
+	tmpCheckResult, err := handlerFunc(ctx)
+	if err != nil {
+		functionName := "intersectionHandler"
+		if !isAllowed {
+			functionName = "exclusionHandler"
+		}
+		c.logger.Error("Failed to execute", zap.Error(err),
+			zap.String("function", functionName),
+			zap.String("object", tmpResult.Object),
+			zap.String("relation", req.Relation),
+			zap.String("user", req.User.String()))
+		return err
+	}
+
+	// If the allowed value does not match what we expect, we skip this candidate.
+	// eg, for intersection we expect the check result to be true
+	// and for exclusion we expect the check result to be false.
+	if tmpCheckResult.GetAllowed() != isAllowed {
+		return nil
+	}
+
+	// If the original stack only had 1 value, we can trySendCandidate right away (nothing more to check)
+	if stack.Len(req.relationStack) == 0 {
+		c.trySendCandidate(ctx, false, tmpResult.Object, resultChan)
+		return nil
+	}
+
+	// If the original stack had more than 1 value, we need to query the parent values
+	// new stack with top item in stack
+	err = c.queryForTuples(ctx, req, false, resultChan, tmpResult.Object)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
 // callCheckForCandidates calls check on the list objects candidates against non lowest weight edges.
 func (c *ReverseExpandQuery) callCheckForCandidates(
 	pool *concurrency.Pool,
@@ -594,48 +616,17 @@ func (c *ReverseExpandQuery) callCheckForCandidates(
 	isAllowed bool,
 ) {
 	pool.Go(func(ctx context.Context) error {
+		// note that we create a separate goroutine pool instead of the main pool
+		// to avoid starvation on the main pool as there could be many candidates
+		// arriving concurrently.
+		tmpResultPool := concurrency.NewPool(ctx, int(c.resolveNodeBreadthLimit))
+
 		for tmpResult := range tmpResultChan {
-			handlerFunc := c.localCheckResolver.CheckRewrite(ctx,
-				&graph.ResolveCheckRequest{
-					StoreID:              req.StoreID,
-					AuthorizationModelID: c.typesystem.GetAuthorizationModelID(),
-					TupleKey:             tuple.NewTupleKey(tmpResult.Object, req.Relation, req.User.String()),
-					ContextualTuples:     req.ContextualTuples,
-					Context:              req.Context,
-					Consistency:          req.Consistency,
-					RequestMetadata:      graph.NewCheckRequestMetadata(),
-				}, userset)
-			tmpCheckResult, err := handlerFunc(ctx)
-			if err != nil {
-				return fmt.Errorf("%w: object: %s, relation: %s, user: %s",
-					ErrCallCheckFail,
-					tmpResult.Object,
-					req.Relation,
-					req.User.String(),
-				)
-			}
-
-			// If the allowed value does not match what we expect, we skip this candidate.
-			// eg, for intersection we expect the check result to be true
-			// and for exclusion we expect the check result to be false.
-			if tmpCheckResult.GetAllowed() != isAllowed {
-				continue
-			}
-
-			// If the original stack only had 1 value, we can trySendCandidate right away (nothing more to check)
-			if stack.Len(req.relationStack) == 0 {
-				c.trySendCandidate(ctx, false, tmpResult.Object, resultChan)
-				continue
-			}
-
-			// If the original stack had more than 1 value, we need to query the parent values
-			// new stack with top item in stack
-			err = c.queryForTuples(ctx, req, false, resultChan, tmpResult.Object)
-			if err != nil {
-				return err
-			}
+			tmpResultPool.Go(func(ctx context.Context) error {
+				return c.callCheckForCandidate(ctx, req, tmpResult, resultChan, userset, isAllowed)
+			})
 		}
-		return nil
+		return tmpResultPool.Wait()
 	})
 }
 
@@ -646,7 +637,7 @@ func (c *ReverseExpandQuery) callCheckForCandidates(
 // resultChan. If check returns false, then these list object candidates
 // are invalid because it does not satisfy all paths for intersection.
 func (c *ReverseExpandQuery) intersectionHandler(
-	ctx context.Context,
+	pool *concurrency.Pool,
 	req *ReverseExpandRequest,
 	resultChan chan<- *ReverseExpandResult,
 	edges []*weightedGraph.WeightedAuthorizationModelEdge,
@@ -707,11 +698,10 @@ func (c *ReverseExpandQuery) intersectionHandler(
 			}}}
 
 	// Concurrently find candidates and call check on them as they are found
-	pool := concurrency.NewPool(ctx, 2)
 	c.findCandidatesForLowestWeightEdge(pool, req, tmpResultChan, lowestWeightEdges, sourceUserType, resolutionMetadata)
 	c.callCheckForCandidates(pool, req, tmpResultChan, resultChan, userset, true)
 
-	return pool.Wait()
+	return nil
 }
 
 // invoke loopOverWeightedEdges to get list objects candidate. Check
@@ -722,6 +712,7 @@ func (c *ReverseExpandQuery) intersectionHandler(
 // are invalid because it does not satisfy all paths for exclusion.
 func (c *ReverseExpandQuery) exclusionHandler(
 	ctx context.Context,
+	pool *concurrency.Pool,
 	req *ReverseExpandRequest,
 	resultChan chan<- *ReverseExpandResult,
 	edges []*weightedGraph.WeightedAuthorizationModelEdge,
@@ -771,9 +762,8 @@ func (c *ReverseExpandQuery) exclusionHandler(
 	}
 
 	// Concurrently find candidates and call check on them as they are found
-	pool := concurrency.NewPool(ctx, 2)
 	c.findCandidatesForLowestWeightEdge(pool, req, tmpResultChan, baseEdges, sourceUserType, resolutionMetadata)
 	c.callCheckForCandidates(pool, req, tmpResultChan, resultChan, userset, false)
 
-	return pool.Wait()
+	return nil
 }