+++ pkg/server/commands/reverseexpand/reverse_expand_weighted_test.go
@@ -12,6 +12,7 @@ import (
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
+	"github.com/openfga/openfga/internal/concurrency"
 	"github.com/openfga/openfga/internal/graph"
 	"github.com/openfga/openfga/internal/mocks"
 	"github.com/openfga/openfga/internal/stack"
@@ -1707,6 +1708,35 @@ func TestReverseExpandWithWeightedGraph(t *testing.T) {
 			expectedOptimizedObjects:   []string{"document:a"},
 			expectedUnoptimizedObjects: []string{"document:a"},
 		},
+		{
+			name: "duplicate_parent_ttu",
+			model: `
+				model
+					schema 1.1
+				type user
+				type thing
+					relations
+						define account: [account]
+						define parent: [account]
+						define can_view: super_admin from account or super_admin from parent
+				type account
+					relations
+						define super_admin: [user]
+		`,
+			tuples: []string{
+				"thing:4#parent@account:4",
+				"account:4#super_admin@user:1",
+			},
+			objectType: "thing",
+			relation:   "can_view",
+			user:       &UserRefObject{Object: &openfgav1.Object{Type: "user", Id: "1"}},
+			expectedOptimizedObjects: []string{
+				"thing:4",
+			},
+			expectedUnoptimizedObjects: []string{
+				"thing:4",
+			},
+		},
 		// TODO: add these when optimization supports infinite weight
 		// intersection with ttu recursive
 		// intersection with userset recursive
@@ -2106,15 +2136,18 @@ func TestIntersectionHandler(t *testing.T) {
 		edges, err := typesys2.GetEdgesFromNode(node, "user")
 		require.NoError(t, err)
 
-		newErr := q.intersectionHandler(ctx, &ReverseExpandRequest{
+		pool := concurrency.NewPool(ctx, 2)
+		err = q.intersectionHandler(pool, &ReverseExpandRequest{
 			StoreID:       storeID,
 			ObjectType:    objectType,
 			Relation:      relation,
 			User:          user,
 			relationStack: nil,
 		}, make(chan *ReverseExpandResult), edges, "", NewResolutionMetadata())
-		require.Error(t, newErr)
-		require.ErrorContains(t, newErr, "invalid edges for source type")
+		require.Error(t, err)
+		require.ErrorContains(t, err, "invalid edges for source type")
+		err = pool.Wait()
+		require.NoError(t, err)
 	})
 
 	t.Run("return_nil_when_there_are_no_connections_for_the_path", func(t *testing.T) {
@@ -2172,8 +2205,10 @@ func TestIntersectionHandler(t *testing.T) {
 		edges, err = typesys.GetEdgesFromNode(edges[0].GetTo(), "user")
 		require.NoError(t, err)
 
+		pool := concurrency.NewPool(ctx, 2)
+
 		go func() {
-			newErr := q.intersectionHandler(ctx, &ReverseExpandRequest{
+			newErr := q.intersectionHandler(pool, &ReverseExpandRequest{
 				StoreID:       storeID,
 				ObjectType:    objectType,
 				Relation:      relation,
@@ -2183,6 +2218,12 @@ func TestIntersectionHandler(t *testing.T) {
 
 			if newErr != nil {
 				errChan <- newErr
+			}
+
+			poolErr := pool.Wait()
+
+			if poolErr != nil {
+				errChan <- poolErr
 			}
 		}()
 
@@ -2253,14 +2294,17 @@ func TestIntersectionHandler(t *testing.T) {
 
 		newStack := stack.Push(nil, typeRelEntry{typeRel: "document#admin"})
 
-		newErr := q.intersectionHandler(ctx, &ReverseExpandRequest{
+		pool := concurrency.NewPool(ctx, 2)
+		err = q.intersectionHandler(pool, &ReverseExpandRequest{
 			StoreID:       storeID,
 			ObjectType:    objectType,
 			Relation:      relation,
 			User:          user,
 			relationStack: newStack,
 		}, make(chan *ReverseExpandResult), edges, "user", NewResolutionMetadata())
-		require.ErrorIs(t, newErr, errorRet)
+		require.NoError(t, err)
+		err = pool.Wait()
+		require.ErrorIs(t, err, errorRet)
 	})
 
 	t.Run("return_error_when_queryForTuples_errors", func(t *testing.T) {
@@ -2320,14 +2364,17 @@ func TestIntersectionHandler(t *testing.T) {
 
 		newStack := stack.Push(nil, typeRelEntry{typeRel: "document#admin"})
 
-		newErr := q.intersectionHandler(ctx, &ReverseExpandRequest{
+		pool := concurrency.NewPool(ctx, 2)
+		err = q.intersectionHandler(pool, &ReverseExpandRequest{
 			StoreID:       storeID,
 			ObjectType:    objectType,
 			Relation:      relation,
 			User:          user,
 			relationStack: newStack,
 		}, make(chan *ReverseExpandResult), edges, "user", NewResolutionMetadata())
-		require.ErrorIs(t, newErr, errorRet)
+		require.NoError(t, err)
+		err = pool.Wait()
+		require.ErrorIs(t, err, errorRet)
 	})
 }
 
@@ -2389,15 +2436,18 @@ func TestExclusionHandler(t *testing.T) {
 		edges, err := typesys2.GetEdgesFromNode(node, "user")
 		require.NoError(t, err)
 
-		newErr := q.exclusionHandler(ctx, &ReverseExpandRequest{
+		pool := concurrency.NewPool(ctx, 2)
+		err = q.exclusionHandler(ctx, pool, &ReverseExpandRequest{
 			StoreID:       storeID,
 			ObjectType:    objectType,
 			Relation:      relation,
 			User:          user,
 			relationStack: nil,
 		}, make(chan *ReverseExpandResult), edges, "", NewResolutionMetadata())
-		require.Error(t, newErr)
-		require.ErrorContains(t, newErr, "invalid exclusion edges for source type")
+		require.Error(t, err)
+		require.ErrorContains(t, err, "invalid exclusion edges for source type")
+		err = pool.Wait()
+		require.NoError(t, err)
 	})
 
 	t.Run("return_error_when_check_errors", func(t *testing.T) {
@@ -2457,14 +2507,17 @@ func TestExclusionHandler(t *testing.T) {
 
 		newStack := stack.Push(nil, typeRelEntry{typeRel: "document#admin"})
 
-		newErr := q.exclusionHandler(ctx, &ReverseExpandRequest{
+		pool := concurrency.NewPool(ctx, 2)
+		err = q.exclusionHandler(ctx, pool, &ReverseExpandRequest{
 			StoreID:       storeID,
 			ObjectType:    objectType,
 			Relation:      relation,
 			User:          user,
 			relationStack: newStack,
 		}, make(chan *ReverseExpandResult), edges, "user", NewResolutionMetadata())
-		require.ErrorIs(t, newErr, errorRet)
+		require.NoError(t, err)
+		err = pool.Wait()
+		require.ErrorIs(t, err, errorRet)
 	})
 
 	t.Run("return_error_when_queryForTuples_errors", func(t *testing.T) {
@@ -2524,13 +2577,16 @@ func TestExclusionHandler(t *testing.T) {
 
 		newStack := stack.Push(nil, typeRelEntry{typeRel: "document#admin"})
 
-		newErr := q.exclusionHandler(ctx, &ReverseExpandRequest{
+		pool := concurrency.NewPool(ctx, 2)
+		err = q.exclusionHandler(ctx, pool, &ReverseExpandRequest{
 			StoreID:       storeID,
 			ObjectType:    objectType,
 			Relation:      relation,
 			User:          user,
 			relationStack: newStack,
 		}, make(chan *ReverseExpandResult), edges, "user", NewResolutionMetadata())
-		require.ErrorIs(t, newErr, errorRet)
+		require.NoError(t, err)
+		err = pool.Wait()
+		require.ErrorIs(t, err, errorRet)
 	})
 }