+++ tests/listobjects/matrix_ttus.go
@@ -304,6 +304,23 @@ var ttus = []matrixTest{
 		},
 	},
 	{
+		Name: "duplicate_ttu_parents",
+		Tuples: []*openfgav1.TupleKey{
+			{Object: "ttus:duplicate_parent", Relation: "mult_parent_types", User: "directs:duplicate_parent"},
+			{Object: "directs:duplicate_parent", Relation: "direct", User: "user:duplicate_parent_anne"},
+		},
+		ListObjectAssertions: []*listobjectstest.Assertion{
+			{
+				Request: &openfgav1.ListObjectsRequest{
+					User:     "user:duplicate_parent_anne",
+					Type:     "ttus",
+					Relation: "duplicate_ttu",
+				},
+				Expectation: []string{"ttus:duplicate_parent"},
+			},
+		},
+	},
+	{
 		Name: "ttus_tuple_cycle_len2_ttu",
 		Tuples: []*openfgav1.TupleKey{
 			// Create a cycle of ttus -> directs